requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
flask_sqlalchemy==3.1.1
pandas==2.2.0
python-dotenv==1.0.1
requests==2.28.2
SQLAlchemy==1.4.42
Werkzeug==2.2.2
